Avance proyecto angular y mejoras en el proyecto backend spring

+++ src/main/java/com/sebastian/supermercado/app/services/impl/ProductoServiceImpl.java
@@ -17,7 +17,7 @@ import com.sebastian.supermercado.app.dto.ProductoDto;
 import com.sebastian.supermercado.app.models.dao.IProductoDao;
 import com.sebastian.supermercado.app.models.entity.Producto;
 import com.sebastian.supermercado.app.models.services.IProductoService;
-import com.sebastian.supermercado.app.response.ProductoResponse;
+import com.sebastian.supermercado.app.response.Response;
 import com.sebastian.supermercado.app.utils.ConverEntitytoDto;
 
 @Service
@@ -32,8 +32,8 @@ public class ProductoServiceImpl implements IProductoService {
 
 	@Override
 	@Transactional(readOnly=true)
-	public ResponseEntity<ProductoResponse> buscarProductos() {
-		ProductoResponse productosResponse = new ProductoResponse();
+	public ResponseEntity<?> buscarProductos() {
+		Response productosResponse = new Response();
 		List<ProductoDto> productos = new ArrayList<>();
 		
 		try {
@@ -41,62 +41,58 @@ public class ProductoServiceImpl implements IProductoService {
 			productos = productosEntity.stream().map(producto -> converEntitytoDto.converToDtoProducto(producto)).collect(Collectors.toList());
 			productosResponse.setProductos(productos);
 		} catch (Exception e) {
-			return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+			return new ResponseEntity<Response>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 		
-		return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.OK);
+		return new ResponseEntity<Response>(productosResponse, HttpStatus.OK);
 	}
 	
 	@Override
 	@Transactional(readOnly=true)
-	public ResponseEntity<ProductoResponse> mostrarProducto(Long id) {
-		ProductoResponse productosResponse = new ProductoResponse();
+	public ResponseEntity<?> mostrarProducto(Long id) {
+		Response productosResponse = new Response();
+		ProductoDto producto = null;
 		
 		try {
-			List<ProductoDto> listaProductos = new ArrayList<>();
 			Producto productoEntity = productoDao.findById(id).orElse(null);
-			ProductoDto producto = converEntitytoDto.converToDtoProducto(productoEntity);
-			listaProductos.add(producto);
-			productosResponse.setProductos(listaProductos);
+			
+			if(productoEntity == null) {
+				log.error("El cliente ID: ".concat(id.toString().concat(" no existe en la base de datos!")));
+				return new ResponseEntity<Response>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+			} else {
+				producto = converEntitytoDto.converToDtoProducto(productoEntity);
+			}
 		} catch (Exception e) {
 			log.error("Error en consultar el producto del supermercado");
-			return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+			return new ResponseEntity<Response>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 		
-		return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.OK);
+		return new ResponseEntity<ProductoDto>(producto, HttpStatus.OK);
 	}
 
 	@Override
 	@Transactional
-	public ResponseEntity<ProductoResponse> crearProducto(ProductoDto producto) {
-		ProductoResponse productosResponse = new ProductoResponse();
-		
+	public ResponseEntity<?> crearProducto(ProductoDto producto) {
 		try {
-			List<ProductoDto> listaProductos = new ArrayList<>();
 			Producto productoEntity = converEntitytoDto.convertToEntityProducto(producto);
 				
 			if(productoEntity != null) {
 				productoDao.save(productoEntity);
-				listaProductos.add(producto);
-				productosResponse.setProductos(listaProductos);
 			} else {
-				log.error("No se puede guardar el producto");
-				return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+				log.error("No se puede guardar un producto nulo");
+				return new ResponseEntity<ProductoDto>(producto, HttpStatus.BAD_REQUEST);  // error 404 : indica que el servidor no puede o no procesará la petición debido a algo que es percibido como un error del cliente
 			}
 		} catch (Exception e) {
-			log.error("Error al guardar el producto");
-			return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+			log.error("Error inesperado en el servidor");
+			return new ResponseEntity<ProductoDto>(producto, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	
-		return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.OK);
+		return new ResponseEntity<ProductoDto>(producto, HttpStatus.OK);
 	}
 	
 	@Override
-	public ResponseEntity<ProductoResponse> guardarProducto(ProductoDto producto, Long id) {
-		ProductoResponse productosResponse = new ProductoResponse();
-		
+	public ResponseEntity<?> guardarProducto(ProductoDto producto, Long id) {
 		try {
-			List<ProductoDto> listaProductos = new ArrayList<>();
 			Producto productoActual = productoDao.findById(id).orElse(null);
 			Producto productoEnviado = converEntitytoDto.convertToEntityProducto(producto);
 			
@@ -108,42 +104,43 @@ public class ProductoServiceImpl implements IProductoService {
 					productoActual.setFechaProducto(new Date());
 					
 					productoDao.save(productoActual);
-					listaProductos.add(producto);
-					productosResponse.setProductos(listaProductos);
 				} else {
 					log.error("error los datos enviados no son correctos para actualizar el producto");
-					return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.BAD_REQUEST);
+					return new ResponseEntity<ProductoDto>(producto, HttpStatus.BAD_REQUEST);
 				}
 			} else {
 				log.error("error no existe el producto que quiere actualizar");
-				return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.NOT_FOUND);
+				return new ResponseEntity<ProductoDto>(producto, HttpStatus.NOT_FOUND);
 			}
 			
 		} catch (Exception e) {
-			log.error("error en actualizar producto");
-			return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+			log.error("Error inesperado en el servidor no se puede actualizar producto");
+			return new ResponseEntity<ProductoDto>(producto, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 		
-		return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.OK);
+		return new ResponseEntity<ProductoDto>(producto, HttpStatus.OK);
 	}
 	
 	@Override
 	@Transactional
-	public ResponseEntity<ProductoResponse> eliminarProducto(Long id) {
-		ProductoResponse productosResponse = new ProductoResponse();
-		List<ProductoDto> listaProductos = new ArrayList<>();
+	public ResponseEntity<?> eliminarProducto(Long id) {
+		ProductoDto productoDto = new ProductoDto();
 		
 		try {
 			Producto productoActual = productoDao.findById(id).orElse(null);
-			ProductoDto productoDto = converEntitytoDto.converToDtoProducto(productoActual);
-			listaProductos.add(productoDto);
-			productosResponse.setProductos(listaProductos);
-			productoDao.deleteById(id);
+			
+			if(productoActual == null) {
+				log.error("error no existe el producto que quiere eliminar");
+				return new ResponseEntity<ProductoDto>(productoDto, HttpStatus.NOT_FOUND);
+			} else {
+				productoDto = converEntitytoDto.converToDtoProducto(productoActual);
+				productoDao.deleteById(id);
+			}
 		} catch (Exception e) {
 			log.error("error al eliminar el producto");
-			return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+			return new ResponseEntity<ProductoDto>(productoDto, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 		
-		return new ResponseEntity<ProductoResponse>(productosResponse, HttpStatus.OK);
+		return new ResponseEntity<ProductoDto>(productoDto, HttpStatus.OK);
 	}
 }
